[MIG] partner_contact_department: Migration to 15.0
