+++ partner_contact_department/readme/CONTRIBUTORS.rst
@@ -1,9 +1,14 @@
-* Pedro M. Baeza <pedro.baeza@tecnativa.com>
-* Antonio Espinosa <antonio.espinosa@tecnativa.com>
-* Rafael Blasco <rafael.blasco@tecnativa.com>
-* Jairo Llopis <jairo.llopis@tecnativa.com>
-* Vicent Cubells <vicent.cubells@tecnativa.com>
-* David Vidal <david.vidal@tecnativa.com>
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+    * Pedro M. Baeza
+    * Antonio Espinosa
+    * Rafael Blasco
+    * Jairo Llopis
+    * Vicent Cubells
+    * David Vidal
+    * Stefan Ungureanu
+
+
 * Nikul Chaudhary <nikulchaudhary2112@gmail.com>
 * Harald Panten <harald.panten@sygel.es>
 * Tharathip Chaweewongphan <tharathipc@ecosoft.co.th>
