+++ partner_contact_department/tests/test_recursion.py
@@ -4,7 +4,7 @@ from odoo.exceptions import UserError
 from odoo.tests import common
 
 
-class TestRecursion(common.SavepointCase):
+class TestRecursion(common.TransactionCase):
     @classmethod
     def setUpClass(cls):
         super(TestRecursion, cls).setUpClass()
