+++ partner_contact_department/views/res_partner_department_view.xml
@@ -9,7 +9,7 @@
         <field name="name">CRM department tree</field>
         <field name="model">res.partner.department</field>
         <field name="arch" type="xml">
-            <tree string="Departments" editable="top">
+            <tree editable="top">
                 <field name="name" />
                 <field name="parent_id" />
             </tree>
